Databricks registrations
