+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/Extensions/DependencyInjection/CalculationResultsRegistration.cs
@@ -12,7 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Azure.Identity;
+using Azure.Storage.Files.DataLake;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Diagnostics.HealthChecks;
 using Energinet.DataHub.Wholesale.CalculationResults.Application.SettlementReports;
 using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResults;
 using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.DataLake;
@@ -20,6 +23,8 @@ using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.JsonSerializ
 using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SettlementReports;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.SettlementReports;
+using Energinet.DataHub.Wholesale.Common.Infrastructure.HealthChecks.DataLake;
+using Energinet.DataHub.Wholesale.Common.Infrastructure.Options;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -37,6 +42,7 @@ public static class CalculationResultsRegistration
         services.AddScoped<ISettlementReportClient, SettlementReportClient>();
 
         services.AddDatabricksSqlStatementExecution(configuration);
+        services.AddDataLakeFileSystemClient(configuration);
 
         services.AddScoped<ISettlementReportResultsCsvWriter, SettlementReportResultsCsvWriter>();
         services.AddScoped<IDataLakeClient, DataLakeClient>();
@@ -51,6 +57,22 @@ public static class CalculationResultsRegistration
                 provider.GetRequiredService<IStreamZipper>()));
         services.AddScoped<ISettlementReportResultQueries, SettlementReportResultQueries>();
 
+        // Health checks
+        services.AddHealthChecks()
+            .AddDatabricksSqlStatementApiHealthCheck()
+            .AddDataLakeHealthCheck(
+                _ => configuration.Get<DataLakeOptions>()!);
+
         return services;
+    }
+
+    private static void AddDataLakeFileSystemClient(this IServiceCollection services, IConfiguration configuration)
+    {
+        var options = configuration.Get<DataLakeOptions>()!;
+        services.AddSingleton<DataLakeFileSystemClient>(_ =>
+        {
+            var dataLakeServiceClient = new DataLakeServiceClient(new Uri(options.STORAGE_ACCOUNT_URI), new DefaultAzureCredential());
+            return dataLakeServiceClient.GetFileSystemClient(options.STORAGE_CONTAINER_NAME);
+        });
     }
 }
