+++ source/dotnet/wholesale-api/Common.Infrastructure/Common.Infrastructure.csproj
@@ -19,6 +19,9 @@ limitations under the License.
     <RootNamespace>Energinet.DataHub.Wholesale.Common.Infrastructure</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="Azure.Identity" Version="1.10.4" />
+    <PackageReference Include="Azure.Storage.Files.DataLake" Version="12.17.1" />
+    <PackageReference Include="Energinet.DataHub.Core.Databricks.Jobs" Version="8.2.1" />
     <PackageReference Include="Energinet.DataHub.Core.App.Common.Abstractions" Version="8.3.1" />
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.22.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="8.0.1" />
