+++ source/dotnet/wholesale-api/Common.Infrastructure/Extensions/DependencyInjection/DatabricksExtensions.cs
@@ -1,0 +1,42 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Energinet.DataHub.Core.Databricks.Jobs.Diagnostics.HealthChecks;
+using Energinet.DataHub.Core.Databricks.Jobs.Extensions.DependencyInjection;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Energinet.DataHub.Wholesale.Common.Infrastructure.Extensions.DependencyInjection
+{
+    /// <summary>
+    /// Extension methods for <see cref="IServiceCollection"/>
+    /// that allow adding Databricks services to an application.
+    /// </summary>
+    public static class DatabricksExtensions
+    {
+        /// <summary>
+        /// Register DatabricksJobs services commonly used by DH3 applications.
+        /// </summary>
+        public static IServiceCollection AddDatabricksJobsForApplication(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddDatabricksJobs(configuration);
+
+            // Health checks
+            services.AddHealthChecks()
+                .AddDatabricksJobsApiHealthCheck();
+
+            return services;
+        }
+    }
+}
