+++ source/dotnet/wholesale-api/Common.Infrastructure/HealthChecks/DataLake/DataLakeHealthRegistration.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.Wholesale.Common.Infrastructure.Options;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using NodaTime;
 
-namespace Energinet.DataHub.Wholesale.WebApi.HealthChecks.DataLake;
+namespace Energinet.DataHub.Wholesale.Common.Infrastructure.HealthChecks.DataLake;
 
 public class DataLakeHealthRegistration : IHealthCheck
 {
