+++ source/dotnet/wholesale-api/WebApi.UnitTests/WebApi/HealthChecks/DataLake/DataLakeHealthCheckTests.cs
@@ -15,8 +15,8 @@
 using Azure;
 using Azure.Storage.Files.DataLake;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
+using Energinet.DataHub.Wholesale.Common.Infrastructure.HealthChecks.DataLake;
 using Energinet.DataHub.Wholesale.Common.Infrastructure.Options;
-using Energinet.DataHub.Wholesale.WebApi.HealthChecks.DataLake;
 using FluentAssertions;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Moq;
@@ -27,6 +27,7 @@ namespace Energinet.DataHub.Wholesale.WebApi.UnitTests.WebApi.HealthChecks.DataL
 
 public class DataLakeHealthCheckTests
 {
+    // TODO: This tests should be moved in another PR (I have written it on the issue)
     [Theory]
     [InlineAutoMoqData(6, 20, 14, false, HealthStatus.Unhealthy)] // Within hour interval but should be Unhealthy because check failed
     [InlineAutoMoqData(6, 20, 14, true, HealthStatus.Healthy)] // Within hour interval should be Healthy because check was successful
